リファクタリング CompanyOutlineViewController
