+++ .gitignore
@@ -48,4 +48,6 @@ DerivedData/
 /CorporateAnalysisHubAPP/GoogleService-Info.plist
 /CorporateAnalysisHubAPP/Secret.swift
 
-Pods
+/Pods
+
+Carthage/
