+++ Cartfile
@@ -1,0 +1,1 @@
+github "xmartlabs/XLPagerTabStrip" ~> 9.0
