+++ CorporateAnalysisHubAPP/ViewController/SearchViewController/SearchResultViewController.swift
